refactor: PEP8 change module name
